added check on storage ref to stop 404 error on delete image when the org is removed

+++ FE/src/app/services/firebase/firebase.service.ts
@@ -133,8 +133,14 @@ export class FirebaseService {
     generalDonationsSummary.delete();
     org.delete()
 
-    this.storage.ref(`Organisations/${orgID}/orgLogo`).delete()
+    const storageRef = this.storage.ref(`Organisations/${orgID}`)
 
+    storageRef.list().forEach((resp) => {
+      console.log(resp.items)
+      if (resp.items.length >= 1) {
+        this.storage.ref(`Organisations/${orgID}/orgLogo`).delete()
+      }
+    })
 
     response = {message: `${orgName} deleted`}
 
@@ -152,7 +158,7 @@ export class FirebaseService {
   async addOrganisation(orgData: any) {
 
     const orgReq:Organisation = {
-      name: orgData.name ? orgData.name : null,
+      name: orgData.name ? orgData.name : "",
       summary: orgData.summary ? orgData.summary: null,
       description: orgData.description ? orgData.description: null,
       activeStatus: orgData.activeStatus ? orgData.activeStatus: null,
